Changed the model class
