+++ Student/Base.lproj/Main.storyboard
@@ -194,7 +194,7 @@
                                 <state key="normal" title="Button"/>
                                 <buttonConfiguration key="configuration" style="filled" title="NEXT"/>
                                 <connections>
-                                    <segue destination="ohh-R4-WFa" kind="presentation" id="sqy-Zu-7k7"/>
+                                    <action selector="findButtontToSearchStudentPagePressed:" destination="BYZ-38-t0r" eventType="touchUpInside" id="1vO-MD-pEx"/>
                                 </connections>
                             </button>
                         </subviews>
@@ -287,6 +287,7 @@
                         <outlet property="idError" destination="naK-KE-5YG" id="XjS-bV-8Kc"/>
                         <outlet property="nameError" destination="fU8-6r-d3d" id="VE6-Pk-GC8"/>
                         <outlet property="nameOfStudent" destination="8H4-0I-cJK" id="4lh-pb-QYo"/>
+                        <outlet property="nextButtonToFindStudentPage" destination="cBs-VR-hDN" id="Mws-Bf-ZaO"/>
                         <outlet property="phoneError" destination="led-jh-KPj" id="dGr-ky-Tbo"/>
                         <outlet property="phoneNo" destination="aTa-r5-YY2" id="TMI-Zr-1dF"/>
                         <outlet property="pinCode" destination="KtG-Pq-gwF" id="PDV-Gg-FBo"/>
@@ -307,7 +308,7 @@
             <objects>
                 <viewController storyboardIdentifier="findStudentViewController" id="ohh-R4-WFa" customClass="findStudentViewController" customModule="Student" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="qdM-n2-o7L">
-                        <rect key="frame" x="0.0" y="0.0" width="390" height="790"/>
+                        <rect key="frame" x="0.0" y="0.0" width="390" height="844"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="FIND THE STUDENT" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="3Hz-Bx-LPm">
